requirements.txt
asgiref==3.5.2
astroid==2.12.13
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
black==22.10.0
click==8.1.3
decorator==5.1.1
dill==0.3.6
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
executing==1.2.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.6.0
isort==5.10.1
jedi==0.18.2
jsonschema==4.17.1
lazy-object-proxy==1.8.0
Markdown==3.4.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.5
pandas==1.5.2
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.13.0
pylint==2.15.6
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
tomlkit==0.11.6
traitlets==5.5.0
uritemplate==4.1.1
wcwidth==0.2.5
wrapt==1.14.1
